+++ torchao/dtypes/floatx/__init__.py
@@ -1,1 +1,2 @@
-from .floatx import FloatxTensorCoreLayout, FloatxTensorCoreAQTTensorImpl, to_scaled_tc_floatx, from_scaled_tc_floatx, _SPLIT_K_MAP
+from .floatx_tensor_core_layout import FloatxTensorCoreLayout, FloatxTensorCoreAQTTensorImpl, to_scaled_tc_floatx, from_scaled_tc_floatx, _SPLIT_K_MAP
+from .float8_layout import Float8AQTTensorImpl, Float8Layout
