+++ torchao/dtypes/uintx/__init__.py
@@ -1,1 +1,6 @@
 from .uintx import UintxTensor, UintxLayout, UintxAQTTensorImpl, to_uintx, _DTYPE_TO_BIT_WIDTH
+from .uint4 import UInt4Tensor
+from .block_sparse_layout import BlockSparseLayout
+from .semi_sparse_layout import SemiSparseLayout
+from .marlin_sparse_layout import MarlinSparseLayout
+from .tensor_core_tiled_layout import TensorCoreTiledLayout
