+++ torchao/dtypes/uintx/uintx.py
@@ -8,7 +8,10 @@ from torchao.dtypes.utils import (
     Layout,
 )
 from torchao.utils import TorchAOBaseTensor
-from torchao.dtypes.affine_quantized_tensor import PlainAQTTensorImpl, register_layout
+from torchao.dtypes.affine_quantized_tensor import (
+    register_layout,
+    PlainAQTTensorImpl
+)
 from torchao.utils import TORCH_VERSION_AT_LEAST_2_3
 
 aten = torch.ops.aten
